requirements.txt

azure-common==1.1.25
azure-core==1.9.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==10.3.0
azure-mgmt-storage==11.2.0
azure-nspkg==3.0.2
azure-storage-blob==12.6.0
azure-storage-common==0.37.1
azure-storage-nspkg==3.1.0
Django==2.1
django-pyodbc-azure==2.1.0.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
docker==4.3.1
Markdown==3.2.2
matplotlib==3.3.1
numpy==1.19.0
opencv-contrib-python==4.4.0.42
opencv-python==4.4.0.42
pandas==1.0.0
Pillow==7.2.0
protobuf==3.13.0
PyJWT==1.7.1
pymongo==3.11.0
pyodbc==4.0.30
requests==2.24.0
scikit-learn==0.23.2
scipy==1.1.0
six==1.15.0
sklearn==0.0
urllib3==1.25.10

